Adds support for Open Spreadsheet format files (*.ods)
Unfortunately for large files existing libraries are not particularly good
at handling them.  large.ods shows a 2Mb ODS file that expands out to 10Mb
and odfpy uses Gb's of RAM and takes over a minute to load it.

Although not particularly pretty this version has judicious use of regex and
XML parsing to pull the relevant information, and can process 26k rows in a
10Mb content.xml in approx 5-7s.