Improve memory footprint.
headers_guess functions reads entire content of file into memory. This leads
to huge memory footprint when working on large excel files. This patch fixes
this problem by reading only first 1000 lines of file into memory. This should
be sufficient to find the header and saves a lot of memory.